deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
numpy 
scikit-learn==1.2.2
Pillow
plotly
numpy==1.26.4